Moving covenant models to the Covenant folder
Files were previously in the Covenant*s* folder

+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/Covenant/ConduitRank.cs
@@ -17,7 +17,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the tier of the conduit rank.
         /// </summary>
         [JsonPropertyName("tier")]
-        public int Tier { get; set; }
+        public long Tier { get; set; }
 
         /// <summary>
         /// Gets the spell tooltip of the conduit rank.
